requirements.txt
playwright==1.42.0
requests==2.31.0
pymongo==4.6.2
python-dotenv==1.0.1
